requirements.txt
Django==4.2.1
djangorestframework==3.14.0
redis==5.2.0
uvicorn==0.21.0
ipdb==0.13.13
